Actividad 8

+++ Actividad_7/src/codigos/funciones.java
@@ -4,7 +4,7 @@ public class funciones {
 
     //metodo para sacar n cantidad de numeros primos devuelva la cadena de los numeros concatenada
     public String primos(int serie) {
-        int [] num_primos = new int[serie];
+        long [] num_primos = new long [serie];
         int primo,cant_primos=0;
         for (int x = 1;cant_primos!=serie;x++){
             primo=0;
@@ -21,7 +21,7 @@ public class funciones {
     }
     //metodo para sacar la secuencia de n cantidad de fibonacci devuelva la cadena de los numeros concatenada
     public String fibonacci(int x){
-        int [] sucesion_fibunacci = new int[x];
+        long [] sucesion_fibunacci = new long[x];
         if (x > 0)
             sucesion_fibunacci[0]=0;
         if (x > 1)
@@ -33,10 +33,10 @@ public class funciones {
         return resultado(sucesion_fibunacci);
     }
     //metodo que concatena el arreglo llenado por los metodos anteriores separandolos con , y saltos de linea cada 15 iteraciones
-    public String resultado(int [] serie) {
+    public String resultado(long [] serie) {
         StringBuilder cadena = new StringBuilder();
         int salto = 0;
-        for (int cont : serie) {
+        for (long cont : serie) {
             cadena.append(cont).append(", ");
             salto++;
             if (salto == 15) {
